fix(console): fix form widget linkage between sign-up identifiers and authn setting checkboxes
