+++ packages/console/src/pages/SignInExperience/PageContent/SignUpAndSignIn/SignUpForm/index.tsx
@@ -1,4 +1,5 @@
-import { useMemo } from 'react';
+import { SignInIdentifier } from '@logto/schemas';
+import { useEffect, useMemo } from 'react';
 import { Controller, useFormContext, useWatch } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 
@@ -17,7 +18,7 @@ import useSignUpPasswordListeners from './use-sign-up-password-listeners';
 function SignUpForm() {
   const { t } = useTranslation(undefined, { keyPrefix: 'admin_console' });
 
-  const { control } = useFormContext<SignInExperienceForm>();
+  const { control, setValue } = useFormContext<SignInExperienceForm>();
 
   const signUpIdentifiers = useWatch({
     control,
@@ -33,6 +34,29 @@ function SignUpForm() {
     () => signUpIdentifiers.length > 0,
     [signUpIdentifiers.length]
   );
+
+  useEffect(() => {
+    if (signUpIdentifiers.length === 0) {
+      setValue('signUp.password', false, { shouldDirty: true });
+    }
+
+    /**
+     * If username is added and as the only identifier, check "Create your password" option.
+     * If username is added but there's already other identifiers, the "Create your password"
+     * option should also be checked, but is handled in the sub-component. @see SignUpIdentifiersEditBox
+     */
+    if (
+      signUpIdentifiers.length === 1 &&
+      signUpIdentifiers[0]?.identifier === SignInIdentifier.Username
+    ) {
+      setValue('signUp.password', true, { shouldDirty: true });
+    }
+
+    const isSignUpVerify = signUpIdentifiers.some(
+      ({ identifier }) => identifier !== SignInIdentifier.Username
+    );
+    setValue('signUp.verify', isSignUpVerify, { shouldDirty: true });
+  }, [setValue, signUpIdentifiers]);
 
   useSignUpPasswordListeners();
